% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{.convert_pandas_series_to_r}
\alias{.convert_pandas_series_to_r}
\title{Convert pandas Series to R vector with proper type handling}
\usage{
.convert_pandas_series_to_r(py_series)
}
\arguments{
\item{py_series}{Python pandas Series}
}
\value{
R vector
}
\description{
Convert pandas Series to R vector with proper type handling
}
\keyword{internal}
