% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{.get_python_type_mappings}
\alias{.get_python_type_mappings}
\title{Get enhanced Python type mappings for R types}
\usage{
.get_python_type_mappings()
}
\value{
Named list of Python types
}
\description{
Get enhanced Python type mappings for R types
}
\keyword{internal}
