% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{.prepare_dataframe_for_python}
\alias{.prepare_dataframe_for_python}
\title{Prepare R data.frame for Python conversion}
\usage{
.prepare_dataframe_for_python(data_frame)
}
\arguments{
\item{data_frame}{R data.frame to prepare}
}
\value{
Prepared data.frame
}
\description{
Prepare R data.frame for Python conversion
}
\keyword{internal}
