% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{enable_sql_debug}
\alias{enable_sql_debug}
\title{Enable SQL debugging}
\usage{
enable_sql_debug(client, datasource, enabled = TRUE, override = list())
}
\arguments{
\item{client}{As returned by \code{\link[=datasource_client]{datasource_client()}}}

\item{datasource}{The name of the datasource}

\item{enabled}{If TRUE (default), enable SQL debugging; if FALSE, disable it}

\item{override}{Configuration values to override (\code{\link[=add_override]{add_override()}})}
}
\value{
Invisible NULL
}
\description{
Enable SQL debugging
}
\examples{
\dontrun{
  client <- datasource_client()
  enable_sql_debug(client, "my_datasource", TRUE)
}
}
\seealso{
\code{\link{query}} for executing raw SQL queries, \code{\link{table_query}} for a fluent query interface
}
