% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{encode_url_path}
\alias{encode_url_path}
\title{Properly escape a URL path}
\usage{
encode_url_path(url)
}
\arguments{
\item{url}{URL to properly escape}
}
\value{
escaped url
}
\description{
Properly escape a URL path
}
\details{
Internal only to fix malformed URLs.
}
