% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{estimate_message_size}
\alias{estimate_message_size}
\title{Estimate message size for a given chunk size}
\usage{
estimate_message_size(
  client,
  datasource,
  data_frame,
  chunk_size,
  override = list()
)
}
\arguments{
\item{client}{As returned by \code{\link[=datasource_client]{datasource_client()}}}

\item{datasource}{The name of the datasource}

\item{data_frame}{Data frame to analyze}

\item{chunk_size}{Number of rows per chunk}

\item{override}{Configuration values to override (\code{\link[=add_override]{add_override()}})}
}
\value{
Numeric value indicating estimated message size in MB
}
\description{
Estimate message size for a given chunk size
}
\examples{
\dontrun{
client <- datasource_client()
df <- data.frame(id = 1:10000, data = rep("large text", 10000))
estimated_size <- estimate_message_size(client, "my_datasource", df, 5000)
cat("Estimated message size for 5000 rows:", estimated_size, "MB\n")
}
}
\seealso{
\code{\link{calculate_optimal_chunk_size}} for calculating optimal chunks
}
