% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{get_db_type}
\alias{get_db_type}
\title{Get database type with enhanced detection}
\usage{
get_db_type(client, datasource, override = list())
}
\arguments{
\item{client}{As returned by \code{\link[=datasource_client]{datasource_client()}}}

\item{datasource}{The name of the datasource}

\item{override}{Configuration values to override (\code{\link[=add_override]{add_override()}})}
}
\value{
Character string indicating database type
}
\description{
Get database type with enhanced detection
}
\examples{
\dontrun{
client <- datasource_client()
db_type <- get_db_type(client, "my_datasource")
cat("Database type:", db_type, "\n")
}
}
\seealso{
\code{\link{set_db_type_override}} for manually setting database type
}
