% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{get_db_type_override}
\alias{get_db_type_override}
\title{Get current database type override}
\usage{
get_db_type_override(client, datasource, override = list())
}
\arguments{
\item{client}{As returned by \code{\link[=datasource_client]{datasource_client()}}}

\item{datasource}{The name of the datasource}

\item{override}{Configuration values to override (\code{\link[=add_override]{add_override()}})}
}
\value{
Character string of current override or NULL if auto-detection is enabled
}
\description{
Get current database type override
}
\examples{
\dontrun{
client <- datasource_client()
current_override <- get_db_type_override(client, "my_datasource")
if (!is.null(current_override)) {
  cat("Database type manually set to:", current_override, "\n")
} else {
  cat("Using auto-detection\n")
}
}
}
\seealso{
\code{\link{set_db_type_override}} for setting database type override
}
