% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netapp_volumes.R
\name{get_file_url}
\alias{get_file_url}
\title{Get URL for a file in a NetApp volume}
\usage{
get_file_url(client, volume_unique_name, file_name)
}
\arguments{
\item{client}{As returned by \code{\link[=netapp_volume_client]{netapp_volume_client()}}}

\item{volume_unique_name}{Unique name of the volume}

\item{file_name}{Name of the file}
}
\value{
URL string for the file
}
\description{
Get URL for a file in a NetApp volume
}
