% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{get_netapp_token}
\alias{get_netapp_token}
\title{Get NetApp volume authentication token}
\usage{
get_netapp_token()
}
\value{
JWT token string or NULL if not available
}
\description{
Get NetApp volume authentication token
}
\details{
Retrieves JWT token for NetApp volume authentication.
Checks DOMINO_TOKEN_FILE and DOMINO_API_PROXY environment variables.
}
\keyword{internal}
