% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{get_supported_db_types}
\alias{get_supported_db_types}
\title{Get supported database types}
\usage{
get_supported_db_types(client, datasource, override = list())
}
\arguments{
\item{client}{As returned by \code{\link[=datasource_client]{datasource_client()}}}

\item{datasource}{The name of the datasource}

\item{override}{Configuration values to override (\code{\link[=add_override]{add_override()}})}
}
\value{
Character vector of supported database types
}
\description{
Get supported database types
}
\examples{
\dontrun{
client <- datasource_client()
supported <- get_supported_db_types(client, "my_datasource")
cat("Supported database types:", paste(supported, collapse = ", "), "\n")
}
}
\seealso{
\code{\link{set_db_type_override}} for setting database type override
}
