% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{get_type_mappings}
\alias{get_type_mappings}
\title{Get type mappings}
\usage{
get_type_mappings(client, datasource, override = list())
}
\arguments{
\item{client}{As returned by \code{\link[=datasource_client]{datasource_client()}}}

\item{datasource}{The name of the datasource}

\item{override}{Configuration values to override (\code{\link[=add_override]{add_override()}})}
}
\value{
A named list of type mappings
}
\description{
Get type mappings
}
\examples{
\dontrun{
  client <- datasource_client()
  mappings <- get_type_mappings(client, "my_datasource")
  print(mappings)
  
  # View specific type mapping
  if ("str" \%in\% names(mappings)) {
    cat("String type maps to:", mappings[["str"]], "\n")
  }
}
}
\seealso{
\code{\link{register_type}} for adding custom type mappings, \code{\link{write_dataframe}} for writing data with custom types
}
