% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netapp_volumes.R
\name{get_volume}
\alias{get_volume}
\title{Get a NetApp volume by name}
\usage{
get_volume(client, name)
}
\arguments{
\item{client}{As returned by \code{\link[=netapp_volume_client]{netapp_volume_client()}}}

\item{name}{Unique name of the volume}
}
\value{
A Volume object
}
\description{
Get a NetApp volume by name
}
