% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netapp_volumes.R
\name{list_files}
\alias{list_files}
\title{List files in a NetApp volume}
\usage{
list_files(client, volume_unique_name, prefix = "", page_size = 1000)
}
\arguments{
\item{client}{As returned by \code{\link[=netapp_volume_client]{netapp_volume_client()}}}

\item{volume_unique_name}{Unique name of the volume}

\item{prefix}{Optional prefix to filter files}

\item{page_size}{Optional number of files to fetch}
}
\value{
A vector of file paths
}
\description{
List files in a NetApp volume
}
