% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netapp_volumes.R
\name{list_snapshots}
\alias{list_snapshots}
\title{List snapshots for a NetApp volume}
\usage{
list_snapshots(client, volume_unique_name, offset = NULL, limit = NULL)
}
\arguments{
\item{client}{As returned by \code{\link[=netapp_volume_client]{netapp_volume_client()}}}

\item{volume_unique_name}{Unique name of the volume}

\item{offset}{Optional offset for pagination}

\item{limit}{Optional limit for pagination}
}
\value{
A list of RemotefsSnapshot objects
}
\description{
List snapshots for a NetApp volume
}
