% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netapp_volumes.R
\name{list_volumes}
\alias{list_volumes}
\title{List NetApp volumes}
\usage{
list_volumes(client, offset = NULL, limit = NULL)
}
\arguments{
\item{client}{As returned by \code{\link[=netapp_volume_client]{netapp_volume_client()}}}

\item{offset}{Optional offset for pagination}

\item{limit}{Optional limit for pagination}
}
\value{
A list of RemotefsVolume objects
}
\description{
List NetApp volumes
}
