% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_oauth_credentials}
\alias{load_oauth_credentials}
\title{Load OAuth Credentials}
\usage{
load_oauth_credentials(env_var, path)
}
\arguments{
\item{env_var}{the environment variable name containing the API proxy URL}

\item{path}{the API endpoint path for retrieving the token}
}
\value{
named list of override configuration values
}
\description{
Load OAuth Credentials
}
