% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http.R
\name{make_request}
\alias{make_request}
\title{Make the HTTP request}
\usage{
make_request(verb, url, config, headers, request_body, write_disk)
}
\arguments{
\item{verb}{A character string containing an HTTP verb}

\item{url}{Signed URL to download object at.}

\item{config}{A list of config values for the REST call.}

\item{headers}{A list of request headers for the REST call.}

\item{request_body}{A character string containing request body data.}

\item{write_disk}{An argument like \code{\link[httr]{write_disk}} to
write the result directly to disk.}
}
\value{
a \code{\link[httr]{response}} object.
}
\description{
Make the HTTP request
}
\details{
Internal only to enable retry for malformed URLs.
}
