% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{netapp_volume_client}
\alias{netapp_volume_client}
\title{Create a client for NetApp volumes}
\usage{
netapp_volume_client(token_file = NULL, token_url = NULL, token = NULL)
}
\arguments{
\item{token_file}{location of file to read token from}

\item{token_url}{url of the location to read the token from}

\item{token}{token to be used to authenticate}
}
\value{
A \code{domino_data.netapp_volumes.NetAppVolumeClient}.
}
\description{
Create a client for NetApp volumes
}
