% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netapp_volumes.R
\name{put_volume_object}
\alias{put_volume_object}
\title{Upload content to a NetApp volume}
\usage{
put_volume_object(client, volume_unique_name, object, what, override = list())
}
\arguments{
\item{client}{As returned by \code{\link[=netapp_volume_client]{netapp_volume_client()}}}

\item{volume_unique_name}{Unique name of the volume}

\item{object}{The object key to create/overwrite}

\item{what}{character vector, raw vector}

\item{override}{Configuration values to override (\code{\link[=add_override]{add_override()}})}
}
\value{
HTTP status message
}
\description{
Upload content to a NetApp volume
}
