% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{register_type}
\alias{register_type}
\title{Register a custom type mapping}
\usage{
register_type(client, datasource, r_type, sql_type, override = list())
}
\arguments{
\item{client}{As returned by \code{\link[=datasource_client]{datasource_client()}}}

\item{datasource}{The name of the datasource}

\item{r_type}{R type (e.g., "Date", "POSIXct", "integer")}

\item{sql_type}{SQL type to map to (e.g., "DATE", "TIMESTAMP", "INTEGER")}

\item{override}{Configuration values to override (\code{\link[=add_override]{add_override()}})}
}
\value{
Invisible NULL
}
\description{
Register a custom type mapping
}
\examples{
\dontrun{
  client <- datasource_client()
  register_type(client, "my_datasource", "Date", "DATE")
}
}
\seealso{
\code{\link{get_type_mappings}} for viewing current type mappings, \code{\link{write_dataframe}} for writing data with custom types
}
