% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netapp_volumes.R
\name{save_volume_object}
\alias{save_volume_object}
\title{Save an object from a NetApp volume to a local file}
\usage{
save_volume_object(
  client,
  volume_unique_name,
  object,
  file = basename(object),
  override = list()
)
}
\arguments{
\item{client}{As returned by \code{\link[=netapp_volume_client]{netapp_volume_client()}}}

\item{volume_unique_name}{Unique name of the volume}

\item{object}{The object to retrieve}

\item{file}{File path to save object at. Defaults to the object base name.}

\item{override}{Configuration values to override (\code{\link[=add_override]{add_override()}})}
}
\value{
File path where object was saved
}
\description{
Save an object from a NetApp volume to a local file
}
