% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{set_db_type_override}
\alias{set_db_type_override}
\title{Set database type override}
\usage{
set_db_type_override(client, datasource, db_type, override = list())
}
\arguments{
\item{client}{As returned by \code{\link[=datasource_client]{datasource_client()}}}

\item{datasource}{The name of the datasource}

\item{db_type}{Database type to force (e.g., 'db2', 'postgresql', 'mysql', 'oracle', 'sqlserver') or NULL to remove override}

\item{override}{Configuration values to override (\code{\link[=add_override]{add_override()}})}
}
\value{
Invisible NULL
}
\description{
Set database type override
}
\examples{
\dontrun{
client <- datasource_client()

# Force DB2 detection
set_db_type_override(client, "my_datasource", "db2")

# Remove override
set_db_type_override(client, "my_datasource", NULL)
}
}
\seealso{
\code{\link{get_db_type}} for getting current database type, \code{\link{get_supported_db_types}} for supported types
}
