% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{set_grpc_message_limits}
\alias{set_grpc_message_limits}
\title{Set gRPC message limits for the datasource}
\usage{
set_grpc_message_limits(
  client,
  datasource,
  max_message_size_mb = 64,
  override = list()
)
}
\arguments{
\item{client}{As returned by \code{\link[=datasource_client]{datasource_client()}}}

\item{datasource}{The name of the datasource}

\item{max_message_size_mb}{Maximum message size in MB (default: 64.0)}

\item{override}{Configuration values to override (\code{\link[=add_override]{add_override()}})}
}
\value{
Invisible NULL
}
\description{
Set gRPC message limits for the datasource
}
\examples{
\dontrun{
client <- datasource_client()
set_grpc_message_limits(client, "my_datasource", max_message_size_mb = 64.0)
}
}
\seealso{
\code{\link{calculate_optimal_chunk_size}} for chunk size optimization
}
