% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{table_exists}
\alias{table_exists}
\title{Check if a table exists in the database}
\usage{
table_exists(client, datasource, table_name, override = list())
}
\arguments{
\item{client}{As returned by \code{\link[=datasource_client]{datasource_client()}}}

\item{datasource}{The name of the datasource}

\item{table_name}{Name of the table to check}

\item{override}{Configuration values to override (\code{\link[=add_override]{add_override()}})}
}
\value{
Logical value indicating if table exists
}
\description{
Check if a table exists in the database
}
\examples{
\dontrun{
client <- datasource_client()
if (table_exists(client, "my_datasource", "my_table")) {
  cat("Table exists\n")
}
}
}
\seealso{
\code{\link{write_dataframe}} for writing data to tables
}
