% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{table_query}
\alias{table_query}
\title{Create a fluent query interface for a table}
\usage{
table_query(client, datasource, table_name, override = list())
}
\arguments{
\item{client}{As returned by \code{\link[=datasource_client]{datasource_client()}}}

\item{datasource}{The name of the datasource to query}

\item{table_name}{Name of the table to query}

\item{override}{Configuration values to override (\code{\link[=add_override]{add_override()}})}
}
\value{
A list with methods for building and executing queries
}
\description{
Create a fluent query interface for a table
}
\examples{
\dontrun{
client <- datasource_client()

# Get all records
all_records <- table_query(client, "my_datasource", "my_table")$all()

# Chained operations
active_records <- table_query(client, "my_datasource", "my_table")$
  filter("is_active = TRUE")$
  select("name, age")$
  order_by("age DESC")$
  all()

# Get first result
first_record <- table_query(client, "my_datasource", "my_table")$first()

# Count results
record_count <- table_query(client, "my_datasource", "my_table")$count()
}
}
\seealso{
\code{\link{query}} for executing raw SQL queries, \code{\link{write_dataframe}} for writing data to tables
}
