% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netapp_volumes.R
\name{upload_volume_object}
\alias{upload_volume_object}
\title{Upload a file to a NetApp volume}
\usage{
upload_volume_object(
  client,
  volume_unique_name,
  object,
  file,
  override = list()
)
}
\arguments{
\item{client}{As returned by \code{\link[=netapp_volume_client]{netapp_volume_client()}}}

\item{volume_unique_name}{Unique name of the volume}

\item{object}{The object key in the volume}

\item{file}{File path to upload}

\item{override}{Configuration values to override (\code{\link[=add_override]{add_override()}})}
}
\value{
HTTP status message
}
\description{
Upload a file to a NetApp volume
}
