% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{drop_table_quietly}
\alias{drop_table_quietly}
\title{Drop a table quietly (suppress errors)}
\usage{
drop_table_quietly(client, datasource, table_name, override = list())
}
\arguments{
\item{client}{As returned by \code{\link[=datasource_client]{datasource_client()}}}

\item{datasource}{The name of the datasource}

\item{table_name}{Name of the table to drop}

\item{override}{Configuration values to override (\code{\link[=add_override]{add_override()}})}
}
\value{
Invisible NULL
}
\description{
Drop a table quietly (suppress errors)
}
\examples{
\dontrun{
  client <- datasource_client()
  drop_table_quietly(client, "my_datasource", "test_table")
}
}
\seealso{
\code{\link{write_dataframe}} for writing data to tables
}
