% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{passthrough_query}
\alias{passthrough_query}
\title{Execute a query with database passthrough wrapper}
\usage{
passthrough_query(client, datasource, query, override = list())
}
\arguments{
\item{client}{As returned by \code{\link[=datasource_client]{datasource_client()}}}

\item{datasource}{The name of the datasource}

\item{query}{SQL query to execute with passthrough}

\item{override}{Configuration values to override (\code{\link[=add_override]{add_override()}})}
}
\value{
Query result
}
\description{
Execute a query with database passthrough wrapper
}
\examples{
\dontrun{
  client <- datasource_client()
  
  # Execute complex query with passthrough
  complex_query <- "SELECT * FROM users u JOIN orders o ON u.id = o.user_id ORDER BY u.created_date"
  result <- passthrough_query(client, "<data_source>", complex_query)
  
  # Use data source-specific functions
  db_specific <- "SELECT user_id, REGEXP_EXTRACT(email, '@(.*)') as domain FROM users"
  result <- passthrough_query(client, "<data_source>", db_specific)
}
}
\seealso{
\code{\link{wrap_passthrough_query}} for getting wrapped query strings, \code{\link{query}} for standard queries, \code{\link{table_query}} for fluent queries
}
