% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{wrap_passthrough_query}
\alias{wrap_passthrough_query}
\title{Wrap a query for database passthrough}
\usage{
wrap_passthrough_query(client, datasource, query, override = list())
}
\arguments{
\item{client}{As returned by \code{\link[=datasource_client]{datasource_client()}}}

\item{datasource}{The name of the datasource}

\item{query}{SQL query to wrap}

\item{override}{Configuration values to override (\code{\link[=add_override]{add_override()}})}
}
\value{
Wrapped query string
}
\description{
Wrap a query for database passthrough
}
\examples{
\dontrun{
  client <- datasource_client()
  
  # Wrap a complex query for inspection
  complex_query <- "SELECT * FROM users u JOIN orders o ON u.id = o.user_id ORDER BY u.created_date"
  wrapped <- wrap_passthrough_query(client, "<data_source>", complex_query)
  print(wrapped)
  
  # Then execute manually
  result <- query(client, "<data_source>", wrapped)
}
}
\seealso{
\code{\link{passthrough_query}} for executing wrapped queries directly, \code{\link{query}} for standard queries
}
